requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django>=4.0,<5.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
# pandas-ta beta naming on PyPI uses 0.3.14b (no trailing 0)
# Fallback to VCS URL to avoid PyPI resolver issues on py311 builders
# Use release tarball to avoid git auth in builders
# GitHub tag uses 0.3.14b0 (with trailing 0); use tarball to avoid git in builders
# Switch to PyPI drop-in that provides the same module name `pandas_ta`
# pandas-ta temporarily removed from server build to avoid py311/numpy resolver conflicts.
# If needed for local analytics, install in the dashboard image or a dev venv.
scipy==1.13.1
scikit-learn==1.5.1
prompt-toolkit==3.0.48
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyarrow==15.0.2
fastparquet>=2023.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
PyYAML==6.0.2

# pydantic for cross-service schema validation
pydantic==2.7.4
httpx==0.27.0
fastjsonschema==2.19.1

streamlit==1.36.0
plotly==5.22.0
ta>=0.11.0

slowapi==0.1.8
aioredis==2.0.1
asyncpg==0.29.0
fakeredis==2.23.2
confluent-kafka==2.3.0
tenacity==8.2.3
# Embeddings and vector search
sentence-transformers==5.1.0
pinecone-client==6.0.0
faiss-cpu==1.12.0  # For GPU acceleration, use faiss-gpu
# metrics
prometheus_client==0.22.1
# MCP2 service dependencies
-r services/mcp2/requirements.txt
